mohqueue: Rename module interface file #82
