+++ src/modules/mohqueue/mohq_db.c
@@ -19,7 +19,7 @@
  *
  */
 
-#include "mohq.h"
+#include "mohqueue_mod.h"
 #include "mohq_db.h"
 #include "mohq_funcs.h"
