+++ src/modules/mohqueue/mohq_funcs.c
@@ -21,7 +21,7 @@
 
 #include <stdarg.h>
 
-#include "mohq.h"
+#include "mohqueue_mod.h"
 #include "mohq_db.h"
 #include "mohq_funcs.h"
